requirements.txt
Django==1.11.13
pytz==2018.9
djangorestframework
